This is a bit worrying, some discrepancies between:
`update_attributes` vs `set`

`set` were mentioned as atomic update, translates to mongodb's `$set`.
Which, in turn bypass typecasting and just put value as it is.

As shown in this test case: The `foo` column, saved as `f`, in the
database, doesn't have any type declared initially. So it then depends
on the first value assigned to it. On this spec, the earlier test cases
assign `String` value, and so the model set `String` as the type for
all `foo`s.

In a way all the saved value should then be a string. But then `set`
comes in, setting the value directly, and set integer at `foo`.

Likely the usage of `set` should be taken care very cautiously.